i8 -> i16

+++ src/physical_move.rs
@@ -46,7 +46,7 @@ impl PhysicalMove {
         }
     }
 
-    pub fn to_sign(&self, board_size:BoardSize, ply:&mut i8) -> String {
+    pub fn to_sign(&self, board_size:BoardSize, ply:&mut i16) -> String {
         match self.address {
             Some(address) => {
                 address.to_physical_sign(board_size)
